+++ paymentservice/src/main/java/com/constellation/paymentservice/requests/PaymentRequest.java
@@ -12,7 +12,7 @@ import java.math.BigDecimal;
 public class PaymentRequest {
     private Integer userId;
     private Integer itemId;
-    private Integer cardNumber;
+    private BigDecimal cardNumber;
     private Integer ccv;
     private String expiryDate;
     private BigDecimal amount;
